+++ src/net/textproto/header.go
@@ -26,8 +26,7 @@ func (h MIMEHeader) Set(key, value string) {
 // It is case insensitive; CanonicalMIMEHeaderKey is used
 // to canonicalize the provided key.
 // If there are no values associated with the key, Get returns "".
-// To access multiple values of a key, or to use non-canonical keys,
-// access the map directly.
+// To use non-canonical keys, access the map directly.
 func (h MIMEHeader) Get(key string) string {
 	if h == nil {
 		return ""
@@ -37,6 +36,18 @@ func (h MIMEHeader) Get(key string) string {
 		return ""
 	}
 	return v[0]
+}
+
+// Values returns all values associated with the given key.
+// It is case insensitive; CanonicalMIMEHeaderKey is
+// used to canonicalize the provided key. To use non-canonical
+// keys, access the map directly.
+// The returned slice is not a copy.
+func (h MIMEHeader) Values(key string) []string {
+	if h == nil {
+		return nil
+	}
+	return h[CanonicalMIMEHeaderKey(key)]
 }
 
 // Del deletes the values associated with key.
